consolidate controllers
